+++ packages/plugin-hexo/src/output.ts
@@ -1,0 +1,33 @@
+import { OutputPlugin } from '@mark-magic/core'
+import path from 'pathe'
+import * as local from '@mark-magic/plugin-local'
+
+export interface Tag {
+  id: string
+  title: string
+}
+
+export function output(options?: { path?: string; base?: string }): OutputPlugin {
+  const root = options?.path ?? path.resolve()
+  const postsPath = path.resolve(root, 'source/_posts')
+  const resourcePath = path.resolve(root, 'source/resources')
+  const p = local.output({
+    rootContentPath: postsPath,
+    rootResourcePath: resourcePath,
+    meta: (it) => ({
+      layout: 'post',
+      title: it.name,
+      abbrlink: it.id,
+      tags: it.extra?.tags.map((it: { title: string }) => it.title),
+      categories: it.path,
+      date: it.created,
+      updated: it.updated,
+    }),
+    contentLink: (it) => path.join('/', options?.base ?? '/', `/p/${it.linkContentId}`),
+    resourceLink: (it) => `/resources/${it.resource.id}${path.extname(it.resource.name)}`,
+    contentPath: (it) => path.resolve(postsPath, it.id + '.md'),
+    resourcePath: (it) => path.resolve(resourcePath, it.id + path.extname(it.name)),
+  })
+  p.name = 'hexo'
+  return p
+}