+++ packages/plugin-docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mark-magic/plugin-docs
 
+## 0.9.0
+
+### Patch Changes
+
+- @mark-magic/core@0.9.0
+- @mark-magic/utils@0.9.0
+- @mark-magic/plugin-local@0.9.0
+
 ## 0.8.0
 
 ### Minor Changes