+++ packages/plugin-joplin/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @mami/plugin-joplin
 
+## 0.9.0
+
 ## 0.8.0
 
 ### Minor Changes